+++ examples/nice_view/run.rs
@@ -1,17 +1,44 @@
+/// this file contain update logic and winit things to boot up a window before Gfx
+use crate::NiceViewScene;
+use ready_paint::{
+    gfx::{Gfx, LimitFPS},
+    Render, RenderEntry,
+};
 use std::sync::Arc;
-use ready_paint::*;
 use winit::{application::ApplicationHandler, dpi::PhysicalSize, window::Window};
+pub async fn run() {
+    let event_loop = winit::event_loop::EventLoop::new().unwrap();
+    event_loop.set_control_flow(winit::event_loop::ControlFlow::Poll);
+    let mut app = App::default();
+    let _ = event_loop.run_app(&mut app);
+}
 
-fn main() {
-    pollster::block_on(async {
-        let event_loop = winit::event_loop::EventLoop::new().unwrap();
-        event_loop.set_control_flow(winit::event_loop::ControlFlow::Poll);
-        let mut app = App::default();
-        let _ = event_loop.run_app(&mut app);
-    })
+#[derive(Default)]
+pub struct App {
+    pub window: Option<Arc<Window>>,
+    pub render: Render,
+    pub first_resize: bool,
 }
 
 impl ApplicationHandler for App {
+    fn about_to_wait(&mut self, _: &winit::event_loop::ActiveEventLoop) {
+        if let RenderEntry::Ready(ref mut gfx) = self.render.entry {
+            let now = std::time::Instant::now();
+            let delta_time = now - gfx.last_update;
+            let time = *gfx.time.lock().unwrap() + delta_time.as_secs_f32();
+            if let LimitFPS::Limit(fps) = gfx.limit_fps {
+                let frame_duration = std::time::Duration::from_secs_f32(1.0 / fps as f32);
+                if delta_time < frame_duration {
+                    spin_sleep::sleep(frame_duration - delta_time);
+                }
+                gfx.last_update = std::time::Instant::now();
+                *gfx.time.lock().unwrap() = time;
+                let real_delta = gfx.last_update - now;
+                gfx.delta_time = real_delta.as_secs_f32();
+                self.window.as_ref().unwrap().request_redraw();
+            }
+        }
+    }
     fn window_event(
         &mut self,
         event_loop: &winit::event_loop::ActiveEventLoop,
@@ -45,7 +72,7 @@ impl ApplicationHandler for App {
             winit::event::WindowEvent::RedrawRequested => {
                 if let RenderEntry::Ready(ref mut gfx) = self.render.entry {
                     if gfx.surface_config.is_some() {
-                        let _ = gfx.test();
+                        self.render.paint();
                     }
                 }
             }
@@ -56,30 +83,20 @@ impl ApplicationHandler for App {
         if let RenderEntry::NotReady = self.render.entry {
             let window = event_loop
                 .create_window(
-                    winit::window::Window::default_attributes()
-                        .with_title("base".to_string()),
+                    winit::window::Window::default_attributes().with_title("nice_view".to_string()),
                 )
                 .unwrap();
-
             let main_owner_window = Arc::new(window);
             self.window = Some(main_owner_window.clone());
             let gfx = pollster::block_on(async move {
-                println!("in async : Loading");
-                let gfx = gfx::Gfx::new(main_owner_window.clone()).await;
-                println!("in async : Ready");
+                println!("in async: Loading");
+                let gfx = Gfx::new(main_owner_window.clone()).await;
+                println!("in async: Ready");
                 return gfx;
             });
-            println!("out of async");
             self.render.entry = RenderEntry::Ready(gfx);
-            println!("resumed: set render entry ready");
+            self.render.add_scene::<NiceViewScene>("check");
             self.first_resize = true;
         }
     }
-}
-#[derive(Default)]
-pub struct App {
-    pub window: Option<Arc<Window>>,
-    pub wgpu_instance: Option<wgpu::Instance>,
-    pub render: Render,
-    pub first_resize: bool,
 }