+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## [Released]
+## [0.1.2] - 2024-01-15
+### Added
+- Added one general usage of all api concepts example
+- enum FPS limit but no control handling instead user customing
+
 ## [0.1.1] - 2024-01-13
 ### Added
 - Added two examples of how to use the library.